requirements.txt
click==8.1.7
anthropic==0.39.0
python-dotenv==1.0.0
rich==13.7.0
